Added minDataRate to the HttpTransport

+++ jetty-server/src/main/java/org/eclipse/jetty/server/HttpConfiguration.java
@@ -65,8 +65,8 @@ public class HttpConfiguration
     private boolean _delayDispatchUntilContent = true;
     private boolean _persistentConnectionsEnabled = true;
     private int _maxErrorDispatches = 10;
-    private long _minRequestDataRate;
-    private long _minResponseDataRate;
+    private int _minRequestDataRate;
+    private int _minResponseDataRate;
     private CookieCompliance _cookieCompliance = CookieCompliance.RFC6265;
     private boolean _notifyRemoteAsyncErrors = true;
 
@@ -486,7 +486,7 @@ public class HttpConfiguration
      * @return The minimum request data rate in bytes per second; or &lt;=0 for no limit (default)
      */
     @ManagedAttribute("The minimum request content data rate in bytes per second")
-    public long getMinRequestDataRate()
+    public int getMinRequestDataRate()
     {
         return _minRequestDataRate;
     }
@@ -494,7 +494,7 @@ public class HttpConfiguration
     /**
      * @param bytesPerSecond The minimum request data rate in bytes per second; or &lt;=0 for no limit (default)
      */
-    public void setMinRequestDataRate(long bytesPerSecond)
+    public void setMinRequestDataRate(int bytesPerSecond)
     {
         _minRequestDataRate=bytesPerSecond;
     }
@@ -503,7 +503,7 @@ public class HttpConfiguration
      * @return The minimum response data rate in bytes per second; or &lt;=0 for no limit (default)
      */
     @ManagedAttribute("The minimum response content data rate in bytes per second")
-    public long getMinResponseDataRate()
+    public int getMinResponseDataRate()
     {
         return _minResponseDataRate;
     }
@@ -511,7 +511,7 @@ public class HttpConfiguration
     /**
      * @param bytesPerSecond The minimum response data rate in bytes per second; or &lt;=0 for no limit (default)
      */
-    public void setMinResponseDataRate(long bytesPerSecond)
+    public void setMinResponseDataRate(int bytesPerSecond)
     {
         _minResponseDataRate=bytesPerSecond;
     }
